Remove Rust wrappers (moved to `skia`) and rename to `skia-sys`.
